Switch include from using <> to "" so that it will search local path instead of global.
Add CMake option specific to Windows that exports all functions to LIB file. From my understanding, this should only be done when BUILD_SHARED_LIBS is set and so no checks needs to be done if it is statically built.

Add NIFTILIB_EXPORT macro to nifticdf for importing global variables. Global variables are not included when using the CMake option and so this must be done manually. The macro is set to empty if building static library or if not on Windows. Otherwise, it is set to __declspec(dllexport/dllimport). dllexport will be set when building the library and dllimport will be used when using the library.

Add NIFTILIB_STATIC macro that needs to be set if wanting to statically link libraries now. If this is not done, then the global variable in nifticdf will be declared using __declspec and errors may occur.

Added CMAKE_MINIMUM_REQUIRED VERSION because its an error when using CMake.